- feature get/display all rules in json format via cli
