+++ .gitignore
@@ -29,3 +29,4 @@ node_modules/
 Thumbs.db
 
 .idea
+/rules.json
